requirements.txt
docling>=2.0.0
docling-core>=2.0.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
langchain-chroma>=0.1.2
langchain-classic
chromadb>=0.5.0
flashrank>=0.2.9
rank-bm25
ollama>=0.4.0
Pillow>=10.0.0
transformers>=4.40.0
sentence-transformers>=3.0.0
gradio
